skillproof-backend/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Caching & Queue
redis==5.0.1
celery==5.3.6

# AI/ML
spacy==3.7.2
scikit-learn==1.4.0
numpy==1.26.3
pandas==2.1.4
sentence-transformers==2.3.1
openai==1.10.0
langchain==0.1.4

# PDF/Document Processing
PyPDF2==3.0.1
python-docx==1.1.0
python-magic==0.4.27

# Explainable AI
shap==0.44.0

# File Storage
boto3==1.34.34

# Email
sendgrid==6.11.0

# Monitoring
sentry-sdk==1.40.0

# Testing
pytest==7.4.4
pytest-cov==4.1.0
pytest-asyncio==0.23.3
httpx==0.26.0

# Code Quality
black==24.1.1
flake8==7.0.0
mypy==1.8.0

# Utilities
python-slugify==8.0.1
pillow==10.2.0
